Move mail theme page to Design section
