+++ src/PrestaShopBundle/Resources/config/routing/admin/improve/design/mail_theme.yml
@@ -1,0 +1,15 @@
+admin_mail_theme_generate_form:
+  path: /
+  methods: [GET]
+  defaults:
+    _controller: PrestaShopBundle:Admin\Improve\Design\MailTheme:index
+    _legacy_controller: AdminMailTheme
+    _legacy_link: AdminMailTheme
+
+admin_mail_theme_generate:
+  path: /generate
+  methods: [POST]
+  defaults:
+    _controller: PrestaShopBundle:Admin\Improve\Design\MailTheme:generateMails
+    _legacy_controller: AdminMailTheme
+    _legacy_link: AdminMailTheme:postGenerateMails
