--- src/PrestaShopBundle/Resources/config/routing/admin/improve/international/mail_theme.yml
@@ -1,17 +1,0 @@
-admin_mail_theme_generate_form:
-  path: /generate
-  methods: [GET]
-  defaults:
-    _controller: PrestaShopBundle:Admin\Improve\International\MailTheme:generateMailsForm
-    _legacy_controller: AdminMailTheme
-    _legacy_link:
-      - AdminMailTheme
-      - AdminMailTheme:generateMails
-
-admin_mail_theme_generate:
-  path: /generate
-  methods: [POST]
-  defaults:
-    _controller: PrestaShopBundle:Admin\Improve\International\MailTheme:generateMails
-    _legacy_controller: AdminMailTheme
-    _legacy_link: AdminMailTheme:postGenerateMails
